Move semver parsing code
