+++ internal/mode/static/telemetry/collector.go
@@ -5,9 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"runtime"
-	"strconv"
 	"strings"
-	"unicode"
 
 	tel "github.com/nginxinc/telemetry-exporter/pkg/telemetry"
 	appsv1 "k8s.io/api/apps/v1"
@@ -298,51 +296,47 @@ func CollectNodeList(ctx context.Context, k8sClient client.Reader) (v1.NodeList,
 	return nodes, nil
 }
 
-// ParseKubeletVersion takes a string and turns it into a semver format.
-func ParseKubeletVersion(s string) (string, error) {
-	s = strings.TrimSpace(s)
-	s = strings.TrimPrefix(s, "v")
+type clusterInformation struct {
+	Platform  string
+	Version   string
+	ClusterID string
+	Nodes     v1.NodeList
+}
 
-	if s == "" {
-		return "", errors.New("string cannot be empty")
+func collectClusterInformation(ctx context.Context, k8sClient client.Reader) (clusterInformation, error) {
+	var clusterInfo clusterInformation
+
+	nodes, err := CollectNodeList(ctx, k8sClient)
+	if err != nil {
+		return clusterInformation{}, fmt.Errorf("failed to collect cluster information: %w", err)
+	}
+	if len(nodes.Items) == 0 {
+		return clusterInformation{}, errors.New("failed to collect cluster information: NodeList length is zero")
 	}
 
-	parts := strings.SplitN(s, ".", 3)
+	clusterInfo.Nodes = nodes
 
-	if _, err := strconv.Atoi(parts[0]); err != nil {
-		return "", errors.New("string must have a number as the major version")
+	var clusterID string
+	if clusterID, err = CollectClusterID(ctx, k8sClient); err != nil {
+		return clusterInformation{}, fmt.Errorf("failed to collect cluster information: %w", err)
+	}
+	clusterInfo.ClusterID = clusterID
+
+	node := nodes.Items[0]
+
+	clusterInfo.Version = "unknown"
+	kubeletVersion := node.Status.NodeInfo.KubeletVersion
+
+	if version, err := parseSemver(kubeletVersion); err == nil {
+		clusterInfo.Version = version
 	}
 
-	if len(parts) == 1 || parts[1] == "" {
-		return "", errors.New("string must have at least a major and minor version specified")
+	var namespaces v1.NamespaceList
+	if err = k8sClient.List(ctx, &namespaces); err != nil {
+		return clusterInformation{}, fmt.Errorf("failed to collect cluster information: %w", err)
 	}
 
-	// in the edge case where the kubeletVersion is missing the patch version and has trailing characters which include
-	// '.' e.g. "v1.27-gke.1067004"
-	if _, err := strconv.Atoi(parts[1]); err != nil && len(parts) == 3 {
-		parts[1] = parts[1] + parts[2]
-		parts = parts[:len(parts)-1]
-	}
+	clusterInfo.Platform = collectK8sPlatform(node, namespaces)
 
-	lastString := parts[len(parts)-1]
-
-	for index := range lastString {
-		// cut off trailing characters after the patch version.
-		// e.g. if kubeletVersion = "1.27.4+500050039", will return "4"
-		if !unicode.IsDigit(rune(lastString[index])) {
-			parts[len(parts)-1] = lastString[:index]
-			break
-		}
-	}
-
-	if len(parts) == 2 {
-		parts = append(parts, "0")
-	}
-
-	// in the case where lastString was ""
-	if parts[2] == "" {
-		parts[2] = "0"
-	}
-
-	return strings.Join(parts, "."), nil
+	return clusterInfo, nil
 }
